Visibility-Collection-Validation/Collectors/Custom-Collector-Python/requirements.txt
pymongo
python-libmaas